+++ simtools/simrun.py
@@ -4,9 +4,11 @@
 Simulation launch services provide the following functionality:
 
 - generating simulation id based on local date and time;
-- generating simulation directory name.
+- generating simulation directory name;
+- creating directory structure for simulation.
 """
 
+import os
 import time
 
 TMP_DIR_PREFIX = "_"
@@ -25,3 +27,14 @@ def generate_sim_dirname(tmp=False, sim_id=None):
     if not sim_id:
         sim_id = generate_sim_id()
     return sim_id if not tmp else TMP_DIR_PREFIX + sim_id
+
+
+def make_dirs(sim_dirname, sim_master_dirname=None, data_dirname=None):
+    """Create directory structure for simulation."""
+    if sim_master_dirname is None:
+        sim_master_dirname = ""
+    if data_dirname is None:
+        data_dirname = ""
+    sim_path = os.path.join(sim_master_dirname, sim_dirname)
+    os.makedirs(os.path.join(sim_path, data_dirname))
+    return sim_path