+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Added saving parameters to a JSON file.
 - Added saving software versions to a JSON file.
 - Added generating simulation id based on local date and time.
+- Added creating directory structure for simulation.
 - Added hierarchy of SimTools exceptions.
 - Added package installer.
 - Added package description.